move each test class to individual file
